# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Rest API Для управления базо данных заданий",
        "title": "Задание",
        "contact": {
            "name": "Artem Vetoshkin",
            "email": "vet_v2002@mail.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/quest": {
            "post": {
                "description": "Добавляет задание включая его название(уникальное), описание, стоимость и тип. Есть обычная задание, которое выполняется как только вызывается метод, сигнализирующий о выполнении для пользователя задачи. И случайная задача, которая выполняется в с вероятностью 0,5.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Добавление задание.",
                "parameters": [
                    {
                        "description": "Информация о добавляемом фильме",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateQuest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Задание успешно добавлен в базу",
                        "schema": {
                            "$ref": "#/definitions/response.Quest"
                        }
                    },
                    "400": {
                        "description": "В теле запроса ошибка",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    },
                    "409": {
                        "description": "Задача с таким название уже существует",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    }
                }
            }
        },
        "/quest/list": {
            "get": {
                "description": "Позволяет получить список заданий.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение списка заданий.",
                "responses": {
                    "200": {
                        "description": "Список заданий успешно сформирован",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Quest"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    }
                }
            }
        },
        "/quest/{quest_id}": {
            "get": {
                "description": "Позволяет информацию о задании по его id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение задания.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Уникальный идентификатор задания",
                        "name": "quest_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Полученное задание",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Quest"
                            }
                        }
                    },
                    "400": {
                        "description": "В пути запросе ошибка",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    },
                    "404": {
                        "description": "Задание с указанным id не найдено",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные об задании. Все переданные поля будут обновлены. Отсутствующие поля будут оставлены без изменений.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Обновление данных об задании.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Уникальный идентификатор задания",
                        "name": "quest_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Информация об обновлении",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateQuest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задание успешно обновлено в базе",
                        "schema": {
                            "$ref": "#/definitions/response.Quest"
                        }
                    },
                    "400": {
                        "description": "В теле запроса ошибка",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    },
                    "404": {
                        "description": "Задание с указанным id не найден",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет информацию о задании из системы по его id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Удаление задания.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Уникальный идентификатор задания",
                        "name": "quest_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задание успешно удалено"
                    },
                    "400": {
                        "description": "В пути запроса ошибка",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    },
                    "404": {
                        "description": "Задание с указанным id не найдено",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/operate.ModelError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "operate.ModelError": {
            "type": "object",
            "properties": {
                "error_message": {
                    "type": "string"
                }
            }
        },
        "request.CreateQuest": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "integer",
                    "format": "uint8",
                    "maximum": 1000,
                    "minimum": 0,
                    "example": 9
                },
                "description": {
                    "type": "string",
                    "example": "Random quest"
                },
                "name": {
                    "type": "string",
                    "example": "Task"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "usual",
                        "random"
                    ],
                    "example": "random"
                }
            }
        },
        "request.UpdateQuest": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "integer",
                    "format": "uint8",
                    "maximum": 1000,
                    "minimum": 0,
                    "example": 9
                },
                "description": {
                    "type": "string",
                    "example": "Random quest"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "usual",
                        "random"
                    ],
                    "example": "random"
                }
            }
        },
...
        "response.Quest": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "integer",
                    "format": "uint8",
                    "maximum": 1000,
                    "minimum": 0,
                    "example": 9
                },
                "description": {
                    "type": "string",
                    "example": "Random quest"
                },
                "id": {
                    "type": "integer",
                    "format": "uint64",
                    "example": 5
                },
                "name": {
                    "type": "string",
                    "example": "Task"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "usual",
                        "random"
                    ],
                    "example": "random"
                }
            }
        },
...
    }
}